refactor(db): extract createManagedLookupCache + add mapBoolColumn helper
Move the generic cache machinery (RefreshingLookupCache, ManagedLookupCacheOptions,
ManagedLookupCache, createManagedLookupCache) into src/db/lookupCache.ts so it is
independently readable, testable, and reusable by future cache consumers.

Add a mapBoolColumn(value) helper in db.ts that collapses the 14 copies of
  Buffer.isBuffer(x) ? x[0] === 1 : x == 1
that were scattered across every row-mapping function into a single call site.

All types are re-exported from db.ts so the public API is unchanged.
